React Essentials: Storing Components in Files & Using a Good Project Structure.
