+++ src/components/Header.jsx
@@ -1,0 +1,21 @@
+import reactImg from "../assets/react-core-concepts.png";
+const reactDescritons = ["Fundamental", "Crucial", "Core"];
+
+function genRandomInt(max) {
+  return Math.floor(Math.random() * (max + 1));
+}
+
+export default function Header() {
+  const description = reactDescritons[genRandomInt(2)];
+
+  return (
+    <header>
+      <img src={reactImg} alt="Stylized atom" />
+      <h1>React Essentials</h1>
+      <p>
+        {description} React concepts you will need for almost any app you are
+        going to build!
+      </p>
+    </header>
+  );
+}
